clona ai

+++ vigilancia/cppProject/driverVigilancia/include/Comunicacao.h
@@ -50,10 +50,10 @@ public:
 
 	//verifica alteração no acel
 
-	short verificaAcel(short x, short y, short z);
+	int verificaAcel(short x, short y, short z);
 
 	//verifica alteração no gyro
-	short verificaGiro(short x, short y, short z);
+	int verificaGiro(short x, short y, short z);
 
 	//Captura o primeiro valor de entrada
 	void setPrimeiroAcel(short X, short Y, short Z);
